requirements.txt
streamlit
pandas 
python-jobspy
groq
numpy
graphviz